requirements.txt
alembic==1.14.1
annotated-types==0.7.0
anyio==4.5.2
async-timeout==5.0.1
asyncpg==0.30.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
dnspython==2.6.1
ecdsa==0.19.0
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.115.11
fastapi-jwt-auth==0.5.0
google-ai-generativelanguage==0.1.0
google-api-core==2.24.1
google-auth==2.38.0
google-cloud-core==2.4.2
google-cloud-storage==3.1.0
google-crc32c==1.6.0.dev2
google-generativeai==0.1.0rc1
google-resumable-media==2.7.2
googleapis-common-protos==1.69.0
greenlet==3.1.1
grpcio==1.70.0
grpcio-status==1.63.0rc1
h11==0.14.0
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.4.5
Mako==1.3.9
MarkupSafe==2.1.5
proto-plus==1.26.0
protobuf==4.25.6
psycopg2-binary==2.9.10
pyasn1==0.4.8
pyasn1_modules==0.4.1
pydantic==1.10.21
pydantic_core==2.27.2
PyJWT==1.7.1
python-dotenv==1.0.1
python-jose==3.4.0
python-magic==0.4.27
python-multipart==0.0.20
requests==2.32.3
rsa==4.9
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.38
SQLAlchemy-Utils==0.41.2
starlette==0.44.0
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.33.0
Werkzeug==3.0.6
zipp==3.20.2
